[LP-DATA] Separation of Lp Registration and dVPN Registration (#7112)
* moving registration client away

* crate reorganise
