+++ nym-registration-client/src/clients/mod.rs
@@ -1,8 +1,6 @@
 // Copyright 2026 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-#![allow(unused)]
-
 pub mod lp;
 pub mod mixnet;
