+++ nym-registration-client/src/error.rs
@@ -80,7 +80,7 @@ pub enum RegistrationClientError {
         gateway_id: String,
         lp_address: std::net::SocketAddr,
         #[source]
-        source: Box<crate::lp_client::LpClientError>,
+        source: Box<crate::LpClientError>,
     },
 
     #[error("failed to register LP with exit gateway {gateway_id} at {lp_address}: {source}")]
@@ -88,7 +88,7 @@ pub enum RegistrationClientError {
         gateway_id: String,
         lp_address: std::net::SocketAddr,
         #[source]
-        source: Box<crate::lp_client::LpClientError>,
+        source: Box<crate::LpClientError>,
     },
 
     #[error("failed to generate randomness: {0}")]
