--- nym-registration-client/src/lp_client/helpers.rs
@@ -1,95 +1,0 @@
-// Copyright 2026 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
-
-#![allow(dead_code)]
-
-use crate::LpClientError;
-use nym_lp::peer::LpRemotePeer;
-use nym_lp::session::{LpAction, LpInput};
-use nym_lp_data::packet::frame::LpFrameKind;
-use nym_lp_data::packet::{ForwardPacketData, LpFrame};
-use nym_registration_common::{
-    LpRegistrationRequest, LpRegistrationResponse, NymNodeLPInformation,
-};
-use rand010::RngExt;
-
-pub(crate) trait LpFrameSendExt {
-    fn to_lp_frame(&self) -> Result<LpFrame, LpClientError>;
-}
-
-pub(crate) trait LpFrameDeliverExt: Sized {
-    fn from_lp_frame(frame: LpFrame) -> Result<Self, LpClientError>;
-}
-
-impl LpFrameSendExt for LpRegistrationRequest {
-    fn to_lp_frame(&self) -> Result<LpFrame, LpClientError> {
-        let request_bytes = self.serialise().map_err(|e| {
-            LpClientError::SendRegistrationRequest(format!("Failed to serialize request: {e}"))
-        })?;
-
-        tracing::debug!(
-            "Sending registration request ({} bytes)",
-            request_bytes.len()
-        );
-
-        Ok(LpFrame::new_registration(request_bytes))
-    }
-}
-
-impl LpFrameDeliverExt for LpRegistrationResponse {
-    fn from_lp_frame(frame: LpFrame) -> Result<Self, LpClientError> {
-        if frame.kind() != LpFrameKind::Registration {
-            return Err(LpClientError::UnexpectedLpPayload { typ: frame.kind() });
-        }
-
-        let response = LpRegistrationResponse::try_deserialise(&frame.content)
-            .map_err(|source| LpClientError::MalformedRegistrationData { source })?;
-
-        Ok(response)
-    }
-}
-
-impl LpFrameSendExt for ForwardPacketData {
-    fn to_lp_frame(&self) -> Result<LpFrame, LpClientError> {
-        let request_bytes = self.to_bytes();
-
-        tracing::trace!(
-            "Sending forward packet data request ({} bytes)",
-            request_bytes.len()
-        );
-
-        Ok(LpFrame::new_forward(request_bytes))
-    }
-}
-
-pub(crate) fn convert_forward_data(request: ForwardPacketData) -> Result<LpInput, LpClientError> {
-    Ok(LpInput::SendFrame(request.to_lp_frame()?))
-}
-
-pub(crate) fn try_convert_forward_response(action: LpAction) -> Result<Vec<u8>, LpClientError> {
-    let response_data = match action {
-        LpAction::DeliverFrame(data) => data,
-        action => return Err(LpClientError::UnexpectedStateMachineAction { action }),
-    };
-
-    if response_data.kind() != LpFrameKind::Forward {
-        return Err(LpClientError::UnexpectedLpPayload {
-            typ: response_data.kind(),
-        });
-    }
-
-    Ok(response_data.content.into())
-}
-
-pub(crate) fn to_lp_remote_peer(data: NymNodeLPInformation) -> LpRemotePeer {
-    LpRemotePeer::new(data.x25519).with_key_digests(data.expected_kem_key_hashes)
-}
-
-pub(crate) async fn exponential_backoff_with_jitter(attempt: u32) {
-    // Exponential backoff with jitter: 100ms, 200ms, 400ms, 800ms, 1600ms (capped)
-    let base_delay_ms = 100u64 * (1 << attempt.min(4));
-    let jitter_ms: u64 = rand010::rng().random_range(0..(base_delay_ms / 4 + 1));
-    let delay = std::time::Duration::from_millis(base_delay_ms + jitter_ms);
-    tracing::info!("Retrying registration after the following delay {delay:?}");
-    tokio::time::sleep(delay).await;
-}
